Fix feature 1 when you order 2 lattes.
When you requested two lattes separately, these were not counted since it was validating that the quantity of a latte was greater than one.

New discountQtt variable in the product class to count the amount of discounted products, initially it is not necessary but in this way you can know how many have been discounted.

New tests is added, although the assertions remain to be added.